Moved Crossover definition to separate file with companion object
